Mostly working.
